project init
I've initialized a pretty standard zig project with nix.

This project isn't meant to be generic and for public use, but rather
is hyperspecific to a problem I need to solve, this is more for making
the code available to people who may need it.
The readme reflects that by being incredibly concise.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>