added support for parsing page numbers
I've written a parser, that keeps parsing until it finds a valid page
number in the form '#page <num>' such as '#page 12' and simply discards
anything that isn't valid. This is as, the error reporting system I am
planning to implement, requires the page number as metadata.

Or actually, perhaps we could make the page number in the error report
optional, and then still generate errors regardless, and only include
page number, if there is a page that these tags belong to.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>