added basic error metadata & fancy reporting
I've added the error metadata struct that gets set upon an error
alongside the actual zig error that is propagated.

I've chosen to only have one zig error, as the handling for each of the
parsing errors is the *same*, and therefore, in the eyes of the program
are the same error, anything else would just introduce unnecessary
complexity.

I've also implemented the fancy reporting to stderr upon errors, with
the context line contents, context span, and error span with arrows
under the important spans, also syntax highlighting.
There's not much for me to write about it, it should be pretty
self-explanitory, the choices I made, I've made as they look pretty.

The only notable thing, is that the context span must contain the error
span or else everything crashes and burns- this is to make my life
easier. If you do not wish to have a context span, then simply set it to
the error span as it will be 'overwritten', but it must be AT LEAST
the same 'size' as the error span.

I've also included some testing code for both the line trimming, the
highlighting and arrows.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>