proper #page parsing error handling & report bugfix
I've implemented some proper error handling to page parsing, so that
now if the tag isn't a page tag it throws an error, and if the number
provided isn't a number then it also throws an error.

I've also improved the error reporting by allowing for the errors spans
to extend beyond the end of the actual contents (for when there's
something missing).

Also changed the unknown page message to grey as it looks nicer, and
differentiates from when the page is known.

Also wrote some demonstration code in main to demonstrate how this would
all work.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>