metadata: added parsing support for metadata
I've implemented parsing support for metadata, that parses one metadata
field at a time, discarding everything inbetween. Currently how it
handles #pages isn't how I want it to be, but I'm writing the program in
a weird way, so that is to be expected.
This is not how I want the code to work, but it does work, so until I do
a proper rewrite, this is a good enough prototype to prod everything out.

Also, I've written some extra parsing for dates in the `yyyy-mm-dd`
format.

And, as per usual, I've written some code in main to demonstrate how
this code *should*, or would work.

The code throws errors, if any of the metadata tags are invalid
(not recognized) and if any of them are missing
(either EOF or the next #page is hit before metadata is complete).
How the metadata tags work, is that you have the tag itself, and
everything following it (stripped of starting and ending whitespace)
until the newline or EOF belong to that metadata tag's contents.
I could just have a counter of fields left to set that decrements for
checking if it's complete or not.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>