modifiers: add paren parsing & error report bugfix
I've added basic parenthesis parsing and also escaping through the use
of a sub-parentheses counter of all the open parentheses. This does
bring the requirement that all content parentheses must also be
complete but you will never see a long `)` in any text.

Also, I've chosen so that modifiers must never span more than one line,
so the program will crash upon a missing parentheses at the end of the
line.

Due to this, I've also discovered a bug in the error reporting where for
contents, it clamps to the end of the contents rather than the line, so
it's possible to include `\n` in the line, which shouldn't happen. Also
added more saturating subs to stop everything crashing, but it seems to
work perfectly.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>