modifier: added general modifier & new page support
I've added some more code that can parse modifiers in the format of
`#red(some red text)`.
The reason why I've chosen to use a writer for formatting, is it just
makes my life easier as it delegates allocating to the calling function
(probably gonna use an ArrayList writer for this).

I've also made it, so that if a #page tag is encountered, it parses the
page and returns the new page number.

This function only be run when a # is encountered whilst converting the
body of the contents.

So far, the only implemented modifier is #title, as that's the simplest
one to implement.
I plan to also add color modifiers and eventually images too.

As per usual, I've also added main code to demo it all.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>