modifiers: bold & itallic
I've decided that, my initial plan to use antiword to generate xml and
then parse that with my custom parser. But at that point, since I'm
going to just use the cli anyways intead of writing bindings to a C++
library, I might as well go the full mile and just strip the text from
the word document and parse it the same as text files.

And due to that, I've decided to just make text effects with modifiers.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>