xml: added basic xml paragraph 'parsing'
I've decided to implement xml parsing for <para> blocks in the xml
that's generated by antiword for word docs. Even though the first
version probably won't use it, it's more of a PoC for the xml parsing.

How this works, is that it finds the first `<para`, assumes it's a
paragraph, then parses everything from the newline to the newline before
the '</' closing tag that corresponds to the `<para` using the same
bracket-matching parsing idea as with parentheses (sub-tag counter).

The main reason for even entertaining the idea of using xml, was for
lists, but now that I think about it, I could just ask them to use the
text vesrion of those lists (- or 1.).

Signed-off-by: Ethan Ma <kalscium@protonmail.com>