word document &amp; 'bugfix'
I've added support for '& < >', as those symbols in word documents are
represented by escape codes of the format &name;.

I've just added some code, so that it checks for these escape codes, if
they're valid and converts them (in the book) if required.

This is just a minor change that is valid for both text and text
generated by antiword versions of the 'source code'.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>