added #cover, full main demo and err report bugfix
The first thing I've done, is implement a #cover tag, which is just a
shortcut for the `#page 0` tag. What I'm planning, is that page 0,
should be the cover page of the logbook, it still has the same
limitations as the other pages, but is included before the table of
contents instead of after.

I've also written a 'full' main demo of the text processing, as it now
iteratively goes through the entire text, updating the page number as
it encounters new pages.

Through this, I also found a flaw with the error reporting system, where
occassionally, the line end is never set and is left on undefined if
it's final value is to be EOF. I've solved it by just initializing it as
EOF.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>